build: upgrade visual snapshot to support new artifacts api (#427)
### Motivation

`getsentry/action-visual-snapshot` has been deprecated and archived, it
is no longer supported, github has recently removed support for the
artifacts API it was using to store the images, causing all of these
actions to fail.

### Modifications

Fork the github action and update it to support the newer github actions
artifacts API.

this is a temporary fix with the intention to replace this github action
with another more supported action in the near term.

### Verification

Checking to see if the build succeeds after this is merged.